feat: complete phase 2 task orchestration model
Finish the structured agent profile and task orchestration slice so new tasks route into durable plans, assignments, and visible state transitions instead of draft-only records.

+++ hq/backend/src/__tests__/task-routing.test.ts
@@ -10,7 +10,7 @@ test('architecture-oriented work routes to software-architect', () => {
 
   assert.equal(decision.recommendedAgentRole, 'software-architect');
   assert.equal(decision.routingStatus, 'matched');
-  assert.equal(decision.executionMode, 'single');
+  assert.equal(decision.executionMode, 'serial');
 });
 
 test('documentation-oriented work routes to technical-writer', () => {
